rename camelcase app files
