+++ app/components/index.ts
@@ -16,7 +16,7 @@ import documentsList from './documents-list';
 import extensionButtons from './extension-buttons';
 import feedback from './feedback';
 import feedbackButton from './feedback-button';
-import helpMarkdown from './helpMarkdown';
+import helpMarkdown from './help-markdown';
 import hiveButton from './hive-button';
 import hivedDocs from './hived-docs';
 import hivers from './hivers';
@@ -38,8 +38,8 @@ import newReply from './new-reply';
 import notFound from './not-found';
 import notifications from './notifications';
 import officeMap from './office-map';
-import passwordRequest from './passwordRequest';
-import passwordReset from './passwordReset';
+import passwordRequest from './password-request';
+import passwordReset from './password-reset';
 import pdf from './pdf';
 import pdfHighlight from './pdf-highlight';
 import phFooter from './ph-footer';
@@ -53,7 +53,7 @@ import signup from './signup';
 import subscribe from './subscribe';
 import subscribed from './subscribed';
 import terms from './terms';
-import urlShare from './urlShare';
+import urlShare from './url-share';
 import user from './user';
 import userProfile from './user-profile';
