+++ app/config/html5-mode.ts
