+++ app/config/index.ts
@@ -1,11 +1,11 @@
 'use strict';
 
 import animate from './animate';
-import html5Mode from './html5Mode';
+import html5Mode from './html5-mode';
 import http from './http';
 import logging from './logging';
 import mathjax from './mathjax';
-import metaUpdate from './metaUpdate';
+import metaUpdate from './meta-update';
 import pdf from './pdf';
 import routes from './routes';
 import scroll from './scroll';
