+++ app/filters/index.ts
@@ -1,5 +1,5 @@
-import highlightsByPageNumber from './highlightsByPageNumber';
-import queryString from './queryString';
+import highlightsByPageNumber from './highlights-by-page-number';
+import queryString from './query-string';
 import range from './range';
 
 export default function(app) {
